Remove flag from ACTION_OPEN_DOCUMENT Intent
FLAG_GRANT_*_URI_PERMISSION is not used to request URI read/write permissions. Those flags are used to signal the OS that the Intent sender wishes to grant the receiver of an Intent the permission to access content: URIs within the Intent. The documentation of ACTION_OPEN_DOCUMENT doesn't mention these flags and unconditionally states:

> All selected documents are returned to the calling application with persistable read and write permission grants.